requirements.txt
# Core Dependencies
google-cloud-vision>=3.7.0
google-cloud-storage>=2.14.0

# Image Processing
Pillow>=10.4.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.31.1
python-multipart>=0.0.9

# Async Processing (optional - comment out if not needed)
celery[redis]>=5.3.6
redis>=5.0.1

# Configuration & Validation
pydantic>=2.11.0
pydantic-settings>=2.5.2
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=24.1.0
prometheus-client>=0.19.0

# CLI
click>=8.1.7
rich>=13.7.0

# Frontend
streamlit>=1.31.0
requests>=2.31.0
streamlit-lottie>=0.0.5

# Testing
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3
httpx>=0.27.1

# Utilities
tenacity>=8.2.3
cachetools>=5.3.2
